refactor: extract shared compile flow and move helpers to common.rs
Move helper functions, MCPG generation, and MCP validation from
standalone.rs to common.rs. Extract compile_shared() function with
CompileConfig struct so both standalone and 1ES compilers can share
the common compilation flow.

Standalone compiler is now a thin wrapper that provides target-specific
values (AWF domains, MCPG config, firewall version) via extra_replacements.

This is a pure refactor - standalone output is byte-identical to before
(verified via golden snapshot comparison).
