+++ src/compile/common.rs
@@ -1,9 +1,11 @@
 //! Common helper functions shared across all compile targets.
 
 use anyhow::{Context, Result};
+use std::collections::{HashMap, HashSet};
+use std::path::Path;
 
 use super::types::{FrontMatter, PipelineParameter, Repository, TriggerConfig};
-use super::extensions::CompilerExtension;
+use super::extensions::{CompilerExtension, Extension, McpgServerConfig, McpgGatewayConfig, McpgConfig, CompileContext};
 use crate::compile::types::McpConfig;
 use crate::fuzzy_schedule;
 
@@ -1218,10 +1220,753 @@ pub fn validate_resolve_pr_thread_statuses(front_matter: &FrontMatter) -> Result
     Ok(())
 }
 
+/// Generate the setup job YAML
+pub fn generate_setup_job(setup_steps: &[serde_yaml::Value], agent_name: &str, pool: &str) -> String {
+    if setup_steps.is_empty() {
+        return String::new();
+    }
+
+    let steps_yaml = format_steps_yaml_indented(setup_steps, 4);
+
+    format!(
+        r#"- job: SetupJob
+  displayName: "{} - Setup"
+  pool:
+    name: {}
+  steps:
+    - checkout: self
+{}
+"#,
+        agent_name, pool, steps_yaml
+    )
+}
+
+/// Generate the teardown job YAML
+pub fn generate_teardown_job(
+    teardown_steps: &[serde_yaml::Value],
+    agent_name: &str,
+    pool: &str,
+) -> String {
+    if teardown_steps.is_empty() {
+        return String::new();
+    }
+
+    let steps_yaml = format_steps_yaml(teardown_steps);
+
+    format!(
+        r#"  - job: TeardownJob
+    displayName: "{} - Teardown"
+    dependsOn: ProcessSafeOutputs
+    pool:
+      name: {}
+    steps:
+      - checkout: self
+{}
+"#,
+        agent_name, pool, steps_yaml
+    )
+}
+
+/// Generate prepare steps (inline), including extension steps and user-defined steps.
+pub fn generate_prepare_steps(
+    prepare_steps: &[serde_yaml::Value],
+    extensions: &[super::extensions::Extension],
+) -> Result<String> {
+    let mut parts = Vec::new();
+
+    // Extension prepare steps and prompt supplements (runtimes + first-party tools)
+    for ext in extensions {
+        for step in ext.prepare_steps() {
+            parts.push(step);
+        }
+        if let Some(prompt) = ext.prompt_supplement() {
+            parts.push(super::extensions::wrap_prompt_append(&prompt, ext.name())?);
+        }
+    }
+
+    if !prepare_steps.is_empty() {
+        parts.push(format_steps_yaml_indented(prepare_steps, 0));
+    }
+
+    Ok(parts.join("\n\n"))
+}
+
+/// Generate finalize steps (inline)
+pub fn generate_finalize_steps(finalize_steps: &[serde_yaml::Value]) -> String {
+    if finalize_steps.is_empty() {
+        return String::new();
+    }
+
+    format_steps_yaml_indented(finalize_steps, 0)
+}
+
+/// Generate dependsOn clause for setup job
+pub fn generate_agentic_depends_on(setup_steps: &[serde_yaml::Value]) -> String {
+    if !setup_steps.is_empty() {
+        "dependsOn: SetupJob".to_string()
+    } else {
+        String::new()
+    }
+}
+
+/// Sensitive host path prefixes that should not be bind-mounted into MCP containers.
+pub const SENSITIVE_MOUNT_PREFIXES: &[&str] = &[
+    "/etc",
+    "/root",
+    "/home",
+    "/proc",
+    "/sys",
+];
+
+/// Docker runtime flag names that grant dangerous host access.
+/// Checked both as `--flag=value` and as `--flag value` (split across two args).
+pub const DANGEROUS_DOCKER_FLAGS: &[&str] = &[
+    "--privileged",
+    "--cap-add",
+    "--security-opt",
+    "--pid",
+    "--network",
+    "--ipc",
+    "--user",
+    "-u",
+    "--add-host",
+    "--entrypoint",
+];
+
+/// Validate a container image name for injection attempts.
+/// Allows `[a-zA-Z0-9./_:-]` which covers standard Docker image references.
+pub fn validate_container_image(image: &str, mcp_name: &str) -> Vec<String> {
+    let mut warnings = Vec::new();
+    if image.is_empty() {
+        warnings.push(format!("Warning: MCP '{}': container image name is empty.", mcp_name));
+        return warnings;
+    }
+    if !image.chars().all(|c| c.is_ascii_alphanumeric() || "._/:-@".contains(c)) {
+        warnings.push(format!(
+            "Warning: MCP '{}': container image '{}' contains unexpected characters. \
+            Image names should only contain [a-zA-Z0-9./_:-@].",
+            mcp_name, image
+        ));
+    }
+    warnings
+}
+
+/// Validate a volume mount source path, warning on sensitive host directories.
+/// Docker socket mounts are escalated to stderr warnings since they grant container escape.
+/// Note: paths are lowercased for comparison to catch cross-platform casing (e.g. `/ETC/shadow`).
+pub fn validate_mount_source(mount: &str, mcp_name: &str) -> Vec<String> {
+    let mut warnings = Vec::new();
+    // Format: "source:dest:mode"
+    if let Some(source) = mount.split(':').next() {
+        let source_lower = source.to_lowercase();
+        if source_lower.contains("docker.sock") {
+            warnings.push(format!(
+                "Warning: MCP '{}': mount '{}' exposes the Docker socket to the MCP container. \
+                This grants full host Docker access and may allow container escape.",
+                mcp_name, mount
+            ));
+            return warnings;
+        }
+        for prefix in SENSITIVE_MOUNT_PREFIXES {
+            // Match exact path or path with trailing separator to avoid false positives
+            // (e.g. /etc matches /etc and /etc/shadow, but not /etc-configs)
+            if source_lower == *prefix || source_lower.starts_with(&format!("{}/", prefix)) {
+                warnings.push(format!(
+                    "Warning: MCP '{}': mount source '{}' references a sensitive host path ({}). \
+                    Ensure this is intentional.",
+                    mcp_name, source, prefix
+                ));
+                break;
+            }
+        }
+    }
+    warnings
+}
+
+/// Validate Docker runtime args for dangerous flags that could escalate privileges.
+/// Also detects volume mounts smuggled via `-v`/`--volume` that bypass `mounts` validation.
+/// Handles both `--flag=value` and `--flag value` (split) forms.
+pub fn validate_docker_args(args: &[String], mcp_name: &str) -> Vec<String> {
+    let mut warnings = Vec::new();
+    for (i, arg) in args.iter().enumerate() {
+        let arg_lower = arg.to_lowercase();
+        // Check for dangerous Docker flags (both --flag=value and --flag value)
+        for dangerous in DANGEROUS_DOCKER_FLAGS {
+            if arg_lower == *dangerous
+                || arg_lower.starts_with(&format!("{}=", dangerous))
+            {
+                let extra_hint = if *dangerous == "--entrypoint" {
+                    " Use the 'entrypoint:' field instead of passing --entrypoint in args."
+                } else {
+                    ""
+                };
+                warnings.push(format!(
+                    "Warning: MCP '{}': Docker arg '{}' grants elevated privileges. \
+                    Ensure this is intentional.{}",
+                    mcp_name, arg, extra_hint
+                ));
+            }
+        }
+        // Check for volume mounts smuggled via args (bypasses mounts validation)
+        if arg == "-v" || arg == "--volume" {
+            if let Some(mount_spec) = args.get(i + 1) {
+                warnings.push(format!(
+                    "Warning: MCP '{}': volume mount '{}' in args bypasses mounts validation. \
+                    Use the 'mounts:' field instead.",
+                    mcp_name, mount_spec
+                ));
+                warnings.extend(validate_mount_source(mount_spec, mcp_name));
+            } else {
+                warnings.push(format!(
+                    "Warning: MCP '{}': '{}' flag is the last arg with no mount spec following it. \
+                    This is likely a malformed args list.",
+                    mcp_name, arg
+                ));
+            }
+        } else if arg_lower.starts_with("-v=") || arg_lower.starts_with("--volume=") {
+            let mount_spec = arg.splitn(2, '=').nth(1).unwrap_or("");
+            warnings.push(format!(
+                "Warning: MCP '{}': volume mount '{}' in args bypasses mounts validation. \
+                Use the 'mounts:' field instead.",
+                mcp_name, mount_spec
+            ));
+            warnings.extend(validate_mount_source(mount_spec, mcp_name));
+        }
+    }
+    warnings
+}
+
+/// Validate that an MCP HTTP URL uses an allowed scheme.
+pub fn validate_mcp_url(url: &str, mcp_name: &str) -> Vec<String> {
+    let mut warnings = Vec::new();
+    if !url.starts_with("https://") && !url.starts_with("http://") {
+        warnings.push(format!(
+            "Warning: MCP '{}': URL '{}' does not use http:// or https:// scheme. \
+            This may not work with MCPG.",
+            mcp_name, url
+        ));
+    }
+    warnings
+}
+
+/// Warn when env values or headers look like they contain inline secrets.
+/// Secrets should use pipeline variables and passthrough ("") instead.
+pub fn warn_potential_secrets(mcp_name: &str, env: &HashMap<String, String>, headers: &HashMap<String, String>) -> Vec<String> {
+    let mut warnings = Vec::new();
+    for (key, value) in env {
+        if !value.is_empty() && (key.to_lowercase().contains("token")
+            || key.to_lowercase().contains("secret")
+            || key.to_lowercase().contains("key")
+            || key.to_lowercase().contains("password")
+            || key.to_lowercase().contains("pat"))
+        {
+            warnings.push(format!(
+                "Warning: MCP '{}': env var '{}' has an inline value that may be a secret. \
+                Use an empty string (\"\") for passthrough from pipeline variables instead.",
+                mcp_name, key
+            ));
+        }
+    }
+    for (key, value) in headers {
+        if value.to_lowercase().contains("bearer ")
+            || key.to_lowercase() == "authorization"
+        {
+            warnings.push(format!(
+                "Warning: MCP '{}': header '{}' may contain inline credentials. \
+                These will appear in plaintext in the compiled pipeline YAML.",
+                mcp_name, key
+            ));
+        }
+    }
+    warnings
+}
+
+/// Validate that a string is a legal environment variable name (`[A-Za-z_][A-Za-z0-9_]*`).
+/// Prevents injection of arbitrary Docker flags via user-controlled front matter keys.
+pub fn is_valid_env_var_name(name: &str) -> bool {
+    let mut chars = name.chars();
+    chars
+        .next()
+        .map_or(false, |c| c.is_ascii_alphabetic() || c == '_')
+        && chars.all(|c| c.is_ascii_alphanumeric() || c == '_')
+}
+
+/// Generate MCPG configuration from front matter.
+///
+/// Converts the front matter `mcp-servers` definitions into MCPG-compatible JSON.
+/// SafeOutputs is always included as an HTTP backend. Extension-contributed MCPG
+/// entries (e.g., azure-devops) are included via the `extensions` parameter.
+pub fn generate_mcpg_config(
+    front_matter: &FrontMatter,
+    ctx: &CompileContext,
+    extensions: &[super::extensions::Extension],
+) -> Result<McpgConfig> {
+    let mut mcp_servers = HashMap::new();
+
+    // SafeOutputs is always included as an HTTP backend.
+    // MCPG runs with --network host, so it reaches SafeOutputs via localhost
+    // (not host.docker.internal, which requires Docker DNS and isn't available
+    // in host network mode on Linux).
+    mcp_servers.insert(
+        "safeoutputs".to_string(),
+        McpgServerConfig {
+            server_type: "http".to_string(),
+            container: None,
+            entrypoint: None,
+            entrypoint_args: None,
+            mounts: None,
+            args: None,
+            url: Some("http://localhost:${SAFE_OUTPUTS_PORT}/mcp".to_string()),
+            headers: Some(HashMap::from([(
+                "Authorization".to_string(),
+                "Bearer ${SAFE_OUTPUTS_API_KEY}".to_string(),
+            )])),
+            env: None,
+            tools: None,
+        },
+    );
+
+    // Add extension-contributed MCPG server entries (e.g., azure-devops)
+    for ext in extensions {
+        for (name, config) in ext.mcpg_servers(ctx)? {
+            mcp_servers.insert(name, config);
+        }
+    }
+
+    for (name, config) in &front_matter.mcp_servers {
+        // Prevent user-defined MCPs from overwriting the reserved safeoutputs backend
+        if name.eq_ignore_ascii_case("safeoutputs") {
+            log::warn!(
+                "MCP name 'safeoutputs' is reserved for the safe outputs HTTP backend — skipping"
+            );
+            continue;
+        }
+
+        // Skip if already auto-configured by an extension (e.g., tools.azure-devops)
+        if mcp_servers.contains_key(name) {
+            continue;
+        }
+
+        let (is_enabled, options) = match config {
+            McpConfig::Enabled(enabled) => (*enabled, None),
+            McpConfig::WithOptions(opts) => (opts.enabled.unwrap_or(true), Some(opts)),
+        };
+
+        if !is_enabled {
+            continue;
+        }
+
+        if let Some(opts) = options {
+            if opts.container.is_some() && opts.url.is_some() {
+                log::warn!(
+                    "MCP '{}': both 'container' and 'url' are set — using 'container' (stdio). \
+                    Remove 'url' to silence this warning.",
+                    name
+                );
+            }
+
+            if let Some(container) = &opts.container {
+                // Container-based stdio MCP (MCPG-native, per spec §3.2.1)
+                for w in validate_container_image(container, name) { eprintln!("{}", w); }
+                // Validate mount paths for sensitive host directories
+                for mount in &opts.mounts {
+                    for w in validate_mount_source(mount, name) { eprintln!("{}", w); }
+                }
+                // Validate Docker runtime args for privilege escalation
+                for w in validate_docker_args(&opts.args, name) { eprintln!("{}", w); }
+                // Warn about potential inline secrets (check headers too in case user set both)
+                for w in warn_potential_secrets(name, &opts.env, &opts.headers) { eprintln!("{}", w); }
+                let entrypoint_args = if opts.entrypoint_args.is_empty() {
+                    None
+                } else {
+                    Some(opts.entrypoint_args.clone())
+                };
+                let args = if opts.args.is_empty() {
+                    None
+                } else {
+                    Some(opts.args.clone())
+                };
+                let mounts = if opts.mounts.is_empty() {
+                    None
+                } else {
+                    Some(opts.mounts.clone())
+                };
+                let env = if opts.env.is_empty() {
+                    None
+                } else {
+                    Some(opts.env.clone())
+                };
+                let tools = if opts.allowed.is_empty() {
+                    None
+                } else {
+                    Some(opts.allowed.clone())
+                };
+                mcp_servers.insert(
+                    name.clone(),
+                    McpgServerConfig {
+                        server_type: "stdio".to_string(),
+                        container: Some(container.clone()),
+                        entrypoint: opts.entrypoint.clone(),
+                        entrypoint_args,
+                        mounts,
+                        args,
+                        url: None,
+                        headers: None,
+                        env,
+                        tools,
+                    },
+                );
+            } else if let Some(url) = &opts.url {
+                // HTTP-based MCP (remote server)
+                for w in validate_mcp_url(url, name) { eprintln!("{}", w); }
+                // Warn about potential inline secrets in headers
+                for w in warn_potential_secrets(name, &HashMap::new(), &opts.headers) { eprintln!("{}", w); }
+                if !opts.env.is_empty() {
+                    eprintln!(
+                        "Warning: MCP '{}': env vars are not supported for HTTP MCPs — they will be ignored. \
+                        Use headers for authentication instead.",
+                        name
+                    );
+                }
+                let headers = if opts.headers.is_empty() {
+                    None
+                } else {
+                    Some(opts.headers.clone())
+                };
+                let tools = if opts.allowed.is_empty() {
+                    None
+                } else {
+                    Some(opts.allowed.clone())
+                };
+                mcp_servers.insert(
+                    name.clone(),
+                    McpgServerConfig {
+                        server_type: "http".to_string(),
+                        container: None,
+                        entrypoint: None,
+                        entrypoint_args: None,
+                        mounts: None,
+                        args: None,
+                        url: Some(url.clone()),
+                        headers,
+                        env: None,
+                        tools,
+                    },
+                );
+            } else {
+                log::warn!("MCP '{}' has no container or url — skipping", name);
+                continue;
+            }
+        } else {
+            log::warn!("MCP '{}' has no container or url — skipping", name);
+        }
+    }
+
+    Ok(McpgConfig {
+        mcp_servers,
+        gateway: McpgGatewayConfig {
+            port: MCPG_PORT,
+            domain: "host.docker.internal".to_string(),
+            api_key: "${MCP_GATEWAY_API_KEY}".to_string(),
+            payload_dir: "/tmp/gh-aw/mcp-payloads".to_string(),
+        },
+    })
+}
+
+/// Generate additional `-e` flags for the MCPG Docker run command.
+///
+/// MCP containers spawned by MCPG may need environment variables that flow from
+/// the pipeline through the MCPG container (passthrough). This function:
+/// 1. Auto-maps `AZURE_DEVOPS_EXT_PAT` from `SC_READ_TOKEN` when `permissions.read` is configured
+/// 2. Collects passthrough env vars (value is `""`) from container-based MCP configs
+///
+/// Only container-based MCPs are considered — HTTP MCPs don't have child containers
+/// that need env passthrough.
+///
+/// Returns flags formatted for inline insertion in the `docker run` command.
+/// The marker sits after the last hardcoded `-e` flag, so the output must
+/// include leading `\\\n` for line continuation when non-empty.
+pub fn generate_mcpg_docker_env(front_matter: &FrontMatter) -> String {
+    let mut env_flags: Vec<String> = Vec::new();
+    let mut seen: HashSet<String> = HashSet::new();
+
+    // Check if any container MCP requests AZURE_DEVOPS_EXT_PAT passthrough
+    let any_mcp_needs_ado_token = front_matter.mcp_servers.values().any(|config| {
+        matches!(config, McpConfig::WithOptions(opts)
+            if opts.enabled.unwrap_or(true)
+                && opts.container.is_some()
+                && opts.env.contains_key("AZURE_DEVOPS_EXT_PAT"))
+    });
+
+    // Also check if tools.azure-devops is enabled (auto-configured ADO MCP always needs token)
+    let ado_tool_needs_token = front_matter
+        .tools
+        .as_ref()
+        .and_then(|t| t.azure_devops.as_ref())
+        .is_some_and(|ado| ado.is_enabled());
+
+    // Auto-map AZURE_DEVOPS_EXT_PAT from SC_READ_TOKEN when permissions.read is configured
+    // AND at least one container MCP requests it via env passthrough (or the ADO tool is enabled)
+    if any_mcp_needs_ado_token || ado_tool_needs_token {
+        if front_matter.permissions.as_ref().and_then(|p| p.read.as_ref()).is_some() {
+            env_flags.push(
+                "-e AZURE_DEVOPS_EXT_PAT=\"$(SC_READ_TOKEN)\"".to_string(),
+            );
+            seen.insert("AZURE_DEVOPS_EXT_PAT".to_string());
+        } else {
+            eprintln!(
+                "Warning: one or more container MCPs request AZURE_DEVOPS_EXT_PAT passthrough \
+                but permissions.read is not configured. The token will be empty at runtime. \
+                Add `permissions: {{ read: <service-connection> }}` to enable auto-mapping."
+            );
+        }
+    }
+
+    // Collect passthrough env vars from container-based MCP configs only.
+    // HTTP MCPs don't have child containers — env passthrough doesn't apply.
+    for (mcp_name, config) in &front_matter.mcp_servers {
+        let opts = match config {
+            McpConfig::WithOptions(opts) if opts.enabled.unwrap_or(true) => opts,
+            _ => continue,
+        };
+
+        // Only container-based MCPs need env passthrough on the MCPG Docker run
+        if opts.container.is_none() {
+            continue;
+        }
+
+        for (var_name, var_value) in &opts.env {
+            // Validate env var name to prevent Docker flag injection (e.g. "X --privileged")
+            if !is_valid_env_var_name(var_name) {
+                log::warn!(
+                    "MCP '{}': skipping invalid env var name '{}' — must match [A-Za-z_][A-Za-z0-9_]*",
+                    mcp_name, var_name
+                );
+                continue;
+            }
+            if seen.contains(var_name) {
+                continue;
+            }
+            // Passthrough: empty string means forward from host/pipeline environment
+            if var_value.is_empty() {
+                env_flags.push(format!("-e {}", var_name));
+                seen.insert(var_name.clone());
+            }
+        }
+    }
+
+    env_flags.sort();
+    if env_flags.is_empty() {
+        // No extra flags — emit a lone `\` so the bash line continuation from the
+        // preceding `-e MCP_GATEWAY_API_KEY=...` flag connects to the image name on
+        // the next line. This is valid bash: a backslash at end-of-line continues
+        // the command. replace_with_indent preserves this on its own indented line.
+        "\\".to_string()
+    } else {
+        // Emit each flag on its own line with `\` continuation.
+        // replace_with_indent handles indentation from the template (base.yml),
+        // so we only emit the content without hardcoded spaces.
+        let flags = env_flags.join(" \\\n");
+        format!("{} \\", flags)
+    }
+}
+
+// ==================== Shared compile flow ====================
+
+/// Target-specific overrides for the shared compile flow.
+pub struct CompileConfig {
+    /// The base YAML template content (the template string itself).
+    pub template: String,
+    /// Additional placeholder→value replacements beyond the shared set.
+    /// These are applied after the shared replacements.
+    pub extra_replacements: Vec<(String, String)>,
+}
+
+/// Shared compilation flow used by both standalone and 1ES compilers.
+///
+/// This function handles the common pipeline compilation steps:
+/// 1. Validates front matter
+/// 2. Generates all shared placeholder values
+/// 3. Runs extension validations
+/// 4. Applies replacements to the template
+/// 5. Prepends the header comment
+///
+/// Target-specific values are provided via `CompileConfig.extra_replacements`.
+pub async fn compile_shared(
+    input_path: &Path,
+    output_path: &Path,
+    front_matter: &FrontMatter,
+    markdown_body: &str,
+    extensions: &[Extension],
+    config: CompileConfig,
+) -> Result<String> {
+    // 1. Validate
+    validate_front_matter_identity(front_matter)?;
+
+    // 2. Generate schedule
+    let schedule = match &front_matter.schedule {
+        Some(s) => generate_schedule(&front_matter.name, s)
+            .with_context(|| format!("Failed to parse schedule '{}'", s.expression()))?,
+        None => String::new(),
+    };
+
+    let repositories = generate_repositories(&front_matter.repositories);
+    let checkout_steps = generate_checkout_steps(&front_matter.checkout);
+    let checkout_self = generate_checkout_self();
+    let agent_name = sanitize_filename(&front_matter.name);
+
+    // 3. Build compile context and run extension validations
+    let input_dir = input_path.parent().unwrap_or(Path::new("."));
+    let ctx = CompileContext::new(front_matter, input_dir).await;
+
+    for ext in extensions {
+        for warning in ext.validate(&ctx)? {
+            eprintln!("Warning: {}", warning);
+        }
+    }
+
+    // 4. Generate copilot params
+    let copilot_params = generate_copilot_params(front_matter, extensions)?;
+
+    // 5. Compute workspace, working directory, triggers
+    let effective_workspace = compute_effective_workspace(
+        &front_matter.workspace,
+        &front_matter.checkout,
+        &front_matter.name,
+    );
+    let working_directory = generate_working_directory(&effective_workspace);
+    let pipeline_resources = generate_pipeline_resources(&front_matter.triggers)?;
+    let has_schedule = front_matter.schedule.is_some();
+    let pr_trigger = generate_pr_trigger(&front_matter.triggers, has_schedule);
+    let ci_trigger = generate_ci_trigger(&front_matter.triggers, has_schedule);
+
+    // 6. Generate source path and pipeline path
+    let source_path = generate_source_path(input_path);
+    let pipeline_path = generate_pipeline_path(output_path);
+
+    // 7. Pool name
+    let pool = front_matter
+        .pool
+        .as_ref()
+        .map(|p| p.name().to_string())
+        .unwrap_or_else(|| DEFAULT_POOL.to_string());
+
+    // 8. Setup/teardown jobs, parameters, prepare/finalize steps
+    let setup_job = generate_setup_job(&front_matter.setup, &front_matter.name, &pool);
+    let teardown_job = generate_teardown_job(&front_matter.teardown, &front_matter.name, &pool);
+    let has_memory = front_matter
+        .tools
+        .as_ref()
+        .and_then(|t| t.cache_memory.as_ref())
+        .is_some_and(|cm| cm.is_enabled());
+    let parameters = build_parameters(&front_matter.parameters, has_memory);
+    let parameters_yaml = generate_parameters(&parameters)?;
+    let prepare_steps = generate_prepare_steps(&front_matter.steps, extensions)?;
+    let finalize_steps = generate_finalize_steps(&front_matter.post_steps);
+    let agentic_depends_on = generate_agentic_depends_on(&front_matter.setup);
+    let job_timeout = generate_job_timeout(front_matter);
+
+    // 9. Token acquisition and env vars
+    let acquire_read_token = generate_acquire_ado_token(
+        front_matter
+            .permissions
+            .as_ref()
+            .and_then(|p| p.read.as_deref()),
+        "SC_READ_TOKEN",
+    );
+    let copilot_ado_env = generate_copilot_ado_env(
+        front_matter
+            .permissions
+            .as_ref()
+            .and_then(|p| p.read.as_deref()),
+    );
+    let acquire_write_token = generate_acquire_ado_token(
+        front_matter
+            .permissions
+            .as_ref()
+            .and_then(|p| p.write.as_deref()),
+        "SC_WRITE_TOKEN",
+    );
+    let executor_ado_env = generate_executor_ado_env(
+        front_matter
+            .permissions
+            .as_ref()
+            .and_then(|p| p.write.as_deref()),
+    );
+
+    // 10. Validations
+    validate_write_permissions(front_matter)?;
+    validate_comment_target(front_matter)?;
+    validate_update_work_item_target(front_matter)?;
+    validate_submit_pr_review_events(front_matter)?;
+    validate_update_pr_votes(front_matter)?;
+    validate_resolve_pr_thread_statuses(front_matter)?;
+
+    // 11. Threat analysis prompt
+    let threat_analysis_prompt = include_str!("../../templates/threat-analysis.md");
+    let template = replace_with_indent(
+        &config.template,
+        "{{ threat_analysis_prompt }}",
+        threat_analysis_prompt,
+    );
+
+    // 12. Shared replacements
+    let compiler_version = env!("CARGO_PKG_VERSION");
+    let replacements: Vec<(&str, &str)> = vec![
+        ("{{ parameters }}", &parameters_yaml),
+        ("{{ compiler_version }}", compiler_version),
+        ("{{ copilot_version }}", COPILOT_CLI_VERSION),
+        ("{{ pool }}", &pool),
+        ("{{ setup_job }}", &setup_job),
+        ("{{ teardown_job }}", &teardown_job),
+        ("{{ prepare_steps }}", &prepare_steps),
+        ("{{ finalize_steps }}", &finalize_steps),
+        ("{{ agentic_depends_on }}", &agentic_depends_on),
+        ("{{ job_timeout }}", &job_timeout),
+        ("{{ repositories }}", &repositories),
+        ("{{ schedule }}", &schedule),
+        ("{{ pipeline_resources }}", &pipeline_resources),
+        ("{{ pr_trigger }}", &pr_trigger),
+        ("{{ ci_trigger }}", &ci_trigger),
+        ("{{ checkout_self }}", &checkout_self),
+        ("{{ checkout_repositories }}", &checkout_steps),
+        ("{{ agent }}", &agent_name),
+        ("{{ agent_name }}", &front_matter.name),
+        ("{{ agent_description }}", &front_matter.description),
+        ("{{ copilot_params }}", &copilot_params),
+        ("{{ source_path }}", &source_path),
+        ("{{ pipeline_path }}", &pipeline_path),
+        ("{{ working_directory }}", &working_directory),
+        ("{{ workspace }}", &working_directory),
+        ("{{ agent_content }}", markdown_body),
+        ("{{ acquire_ado_token }}", &acquire_read_token),
+        ("{{ copilot_ado_env }}", &copilot_ado_env),
+        ("{{ acquire_write_token }}", &acquire_write_token),
+        ("{{ executor_ado_env }}", &executor_ado_env),
+    ];
+
+    let mut pipeline_yaml = replacements
+        .into_iter()
+        .fold(template, |yaml, (placeholder, replacement)| {
+            replace_with_indent(&yaml, placeholder, replacement)
+        });
+
+    // 13. Apply extra replacements (target-specific)
+    for (placeholder, replacement) in &config.extra_replacements {
+        pipeline_yaml = replace_with_indent(&pipeline_yaml, placeholder, replacement);
+    }
+
+    // 14. Prepend header
+    let header = generate_header_comment(input_path);
+    Ok(format!("{}{}", header, pipeline_yaml))
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::compile::types::{McpConfig, McpOptions, Repository};
+    use crate::compile::extensions::{CompileContext, collect_extensions};
+    use std::collections::HashMap;
 
     /// Helper: create a minimal FrontMatter by parsing YAML
     fn minimal_front_matter() -> FrontMatter {
@@ -2569,5 +3314,982 @@ mod tests {
         assert!(result.contains("source: 'Build''s Pipeline'"));
         assert!(result.contains("project: 'My''Project'"));
         assert!(result.contains("- 'it''s-branch'"));
+    }
+
+    // ─── generate_prepare_steps ──────────────────────────────────────────────
+
+    #[test]
+    fn test_generate_prepare_steps_with_memory_includes_memory_preamble() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  cache-memory: true\n---\n",
+        ).unwrap();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let result = generate_prepare_steps(&[], &exts).unwrap();
+        assert!(
+            !result.is_empty(),
+            "memory steps must be emitted when cache-memory enabled"
+        );
+        assert!(
+            result.contains("agent_memory"),
+            "should reference memory directory"
+        );
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_without_memory_and_no_steps_is_empty() {
+        let fm = minimal_front_matter();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let result = generate_prepare_steps(&[], &exts).unwrap();
+        assert!(result.is_empty(), "no steps and no memory should produce empty output");
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_with_memory_includes_download_and_prompt() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  cache-memory: true\n---\n",
+        ).unwrap();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let result = generate_prepare_steps(&[], &exts).unwrap();
+        assert!(
+            result.contains("DownloadPipelineArtifact"),
+            "memory steps must include the artifact download task"
+        );
+        assert!(
+            result.contains("Agent Memory"),
+            "memory steps must include the memory prompt"
+        );
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_without_memory_with_user_steps() {
+        let fm = minimal_front_matter();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let step: serde_yaml::Value =
+            serde_yaml::from_str("bash: echo hello\ndisplayName: greet").unwrap();
+        let result = generate_prepare_steps(&[step], &exts).unwrap();
+        assert!(!result.is_empty(), "user steps should be present");
+        assert!(
+            !result.contains("agent_memory"),
+            "no memory reference when cache-memory not enabled"
+        );
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_with_memory_and_user_steps() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  cache-memory: true\n---\n",
+        ).unwrap();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let step: serde_yaml::Value =
+            serde_yaml::from_str("bash: echo hello\ndisplayName: greet").unwrap();
+        let result = generate_prepare_steps(&[step], &exts).unwrap();
+        assert!(
+            result.contains("agent_memory"),
+            "memory reference must be present"
+        );
+        assert!(
+            result.contains("echo hello"),
+            "user step must also be present"
+        );
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_with_lean() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\nruntimes:\n  lean: true\n---\n",
+        ).unwrap();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let result = generate_prepare_steps(&[], &exts).unwrap();
+        assert!(result.contains("elan-init.sh"), "should include elan installer");
+        assert!(result.contains("Lean 4"), "should include Lean prompt");
+        assert!(result.contains("--default-toolchain stable"), "should default to stable");
+        assert!(result.contains("/tmp/awf-tools/"), "should symlink into awf-tools for AWF chroot");
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_with_lean_custom_toolchain() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\nruntimes:\n  lean:\n    toolchain: \"leanprover/lean4:v4.29.1\"\n---\n",
+        ).unwrap();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let result = generate_prepare_steps(&[], &exts).unwrap();
+        assert!(
+            result.contains("--default-toolchain leanprover/lean4:v4.29.1"),
+            "should use specified toolchain"
+        );
+    }
+
+    #[test]
+    fn test_generate_prepare_steps_with_lean_and_memory() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\nruntimes:\n  lean: true\ntools:\n  cache-memory: true\n---\n",
+        ).unwrap();
+        let exts = crate::compile::extensions::collect_extensions(&fm);
+        let result = generate_prepare_steps(&[], &exts).unwrap();
+        assert!(result.contains("agent_memory"), "memory steps present");
+        assert!(result.contains("elan-init.sh"), "lean install present");
+        assert!(result.contains("Lean 4"), "lean prompt present");
+    }
+
+    // ═══════════════════════════════════════════════════════════════════════
+    // Tests moved from standalone.rs — MCPG config, docker env, validation
+    // ═══════════════════════════════════════════════════════════════════════
+
+    #[test]
+    fn test_generate_firewall_config_custom_mcp() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "my-tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                entrypoint: Some("node".to_string()),
+                entrypoint_args: vec!["server.js".to_string()],
+                allowed: vec!["do_thing".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let server = config.mcp_servers.get("my-tool").unwrap();
+        assert_eq!(server.server_type, "stdio");
+        assert_eq!(server.container.as_ref().unwrap(), "node:20-slim");
+        assert_eq!(server.entrypoint.as_ref().unwrap(), "node");
+        assert_eq!(
+            server.entrypoint_args.as_ref().unwrap(),
+            &vec!["server.js"]
+        );
+        assert_eq!(
+            server.tools.as_ref().unwrap(),
+            &vec!["do_thing".to_string()]
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_mcp_without_transport_skipped() {
+        let mut fm = minimal_front_matter();
+        // An MCP with no container or url should be skipped
+        fm.mcp_servers
+            .insert("phantom".to_string(), McpConfig::Enabled(true));
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        assert!(!config.mcp_servers.contains_key("phantom"));
+        // safeoutputs is always present
+        assert!(config.mcp_servers.contains_key("safeoutputs"));
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_disabled_mcp_skipped() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers
+            .insert("my-tool".to_string(), McpConfig::Enabled(false));
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        assert!(!config.mcp_servers.contains_key("my-tool"));
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_empty_mcp_servers() {
+        let fm = minimal_front_matter();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        // Only safeoutputs should be present
+        assert_eq!(config.mcp_servers.len(), 1);
+        assert!(config.mcp_servers.contains_key("safeoutputs"));
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_gateway_defaults() {
+        let fm = minimal_front_matter();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        assert_eq!(config.gateway.port, 80);
+        assert_eq!(config.gateway.domain, "host.docker.internal");
+        assert_eq!(config.gateway.api_key, "${MCP_GATEWAY_API_KEY}");
+        assert_eq!(config.gateway.payload_dir, "/tmp/gh-aw/mcp-payloads");
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_json_roundtrip() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "my-tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("python:3.12-slim".to_string()),
+                entrypoint: Some("python".to_string()),
+                entrypoint_args: vec!["-m".to_string(), "server".to_string()],
+                allowed: vec!["query".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let json = serde_json::to_string_pretty(&config).expect("Config should serialize to JSON");
+        let parsed: serde_json::Value =
+            serde_json::from_str(&json).expect("Serialized JSON should parse back");
+
+        // Verify top-level structure matches MCPG expectation
+        assert!(
+            parsed.get("mcpServers").is_some(),
+            "Should have mcpServers key"
+        );
+        assert!(parsed.get("gateway").is_some(), "Should have gateway key");
+
+        let gw = parsed.get("gateway").unwrap();
+        assert!(gw.get("port").is_some(), "Gateway should have port");
+        assert!(gw.get("domain").is_some(), "Gateway should have domain");
+        assert!(gw.get("apiKey").is_some(), "Gateway should have apiKey");
+        assert!(
+            gw.get("payloadDir").is_some(),
+            "Gateway should have payloadDir"
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_safeoutputs_variable_placeholders() {
+        let fm = minimal_front_matter();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let so = config.mcp_servers.get("safeoutputs").unwrap();
+
+        // URL should reference the runtime-substituted port
+        let url = so.url.as_ref().unwrap();
+        assert!(
+            url.contains("${SAFE_OUTPUTS_PORT}"),
+            "SafeOutputs URL should use ${{SAFE_OUTPUTS_PORT}} placeholder, got: {url}"
+        );
+
+        // Auth header should reference the runtime-substituted API key
+        let headers = so.headers.as_ref().unwrap();
+        let auth = headers.get("Authorization").unwrap();
+        assert!(
+            auth.contains("${SAFE_OUTPUTS_API_KEY}"),
+            "SafeOutputs auth header should use ${{SAFE_OUTPUTS_API_KEY}} placeholder, got: {auth}"
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_safeoutputs_is_http_type() {
+        let fm = minimal_front_matter();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let so = config.mcp_servers.get("safeoutputs").unwrap();
+        assert_eq!(so.server_type, "http");
+        assert!(
+            so.container.is_none(),
+            "HTTP backend should have no container"
+        );
+        assert!(so.args.is_none(), "HTTP backend should have no args");
+        assert!(so.url.is_some(), "HTTP backend must have a URL");
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_container_mcp_is_stdio_type() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "runner".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                entrypoint: Some("node".to_string()),
+                entrypoint_args: vec!["srv.js".to_string()],
+                allowed: vec!["run".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let srv = config.mcp_servers.get("runner").unwrap();
+        assert_eq!(srv.server_type, "stdio");
+        assert!(
+            srv.container.is_some(),
+            "stdio server must have a container"
+        );
+        assert!(srv.url.is_none(), "stdio server should have no URL");
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_container_with_env() {
+        let mut fm = minimal_front_matter();
+        let mut env = HashMap::new();
+        env.insert("TOKEN".to_string(), "secret".to_string());
+        fm.mcp_servers.insert(
+            "with-env".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                env,
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let srv = config.mcp_servers.get("with-env").unwrap();
+        let e = srv.env.as_ref().unwrap();
+        assert_eq!(e.get("TOKEN").unwrap(), "secret");
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_reserved_safeoutputs_name_rejected() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "safeoutputs".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("evil:latest".to_string()),
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        // The reserved entry should still be the HTTP backend, not the user's container
+        let so = config.mcp_servers.get("safeoutputs").unwrap();
+        assert_eq!(
+            so.server_type, "http",
+            "safeoutputs should remain HTTP backend"
+        );
+        assert!(
+            so.container.is_none(),
+            "User container should not overwrite safeoutputs"
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_safeoutputs_reserved_name_skipped() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "SafeOutputs".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                entrypoint: Some("node".to_string()),
+                entrypoint_args: vec!["evil.js".to_string()],
+                allowed: vec!["hijack".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        // The user-defined "SafeOutputs" must not overwrite the built-in entry
+        let so = config.mcp_servers.get("safeoutputs").unwrap();
+        assert_eq!(so.server_type, "http");
+        assert!(so.url.as_ref().unwrap().contains("localhost"));
+        // No stdio entry should have been added under any casing
+        assert_eq!(config.mcp_servers.len(), 1);
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_http_mcp() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "remote".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                url: Some("https://mcp.example.com/api".to_string()),
+                headers: {
+                    let mut h = HashMap::new();
+                    h.insert("X-Custom".to_string(), "value".to_string());
+                    h
+                },
+                allowed: vec!["query".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let srv = config.mcp_servers.get("remote").unwrap();
+        assert_eq!(srv.server_type, "http");
+        assert_eq!(
+            srv.url.as_ref().unwrap(),
+            "https://mcp.example.com/api"
+        );
+        assert_eq!(
+            srv.headers.as_ref().unwrap().get("X-Custom").unwrap(),
+            "value"
+        );
+        assert!(srv.container.is_none(), "HTTP server should have no container");
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_container_with_entrypoint() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "ado".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                entrypoint: Some("npx".to_string()),
+                entrypoint_args: vec!["-y".to_string(), "@azure-devops/mcp".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let srv = config.mcp_servers.get("ado").unwrap();
+        assert_eq!(srv.server_type, "stdio");
+        assert_eq!(srv.container.as_ref().unwrap(), "node:20-slim");
+        assert_eq!(srv.entrypoint.as_ref().unwrap(), "npx");
+        assert_eq!(
+            srv.entrypoint_args.as_ref().unwrap(),
+            &vec!["-y", "@azure-devops/mcp"]
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_container_with_mounts() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "data-tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("data-tool:latest".to_string()),
+                mounts: vec!["/host/data:/app/data:ro".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let srv = config.mcp_servers.get("data-tool").unwrap();
+        assert_eq!(
+            srv.mounts.as_ref().unwrap(),
+            &vec!["/host/data:/app/data:ro"]
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_config_no_transport_skipped() {
+        let mut fm = minimal_front_matter();
+        // MCP with options but no container or url should be skipped
+        fm.mcp_servers.insert(
+            "no-transport".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                allowed: vec!["tool".to_string()],
+                ..Default::default()
+            }),
+        );
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        assert!(!config.mcp_servers.contains_key("no-transport"));
+    }
+
+    #[test]
+    fn test_generate_mcpg_docker_env_with_permissions_read() {
+        let mut fm = minimal_front_matter();
+        fm.permissions = Some(crate::compile::types::PermissionsConfig {
+            read: Some("my-read-sc".to_string()),
+            write: None,
+        });
+        // A container MCP must request AZURE_DEVOPS_EXT_PAT for the auto-map to trigger
+        fm.mcp_servers.insert(
+            "ado-tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                env: {
+                    let mut e = HashMap::new();
+                    e.insert("AZURE_DEVOPS_EXT_PAT".to_string(), "".to_string());
+                    e
+                },
+                ..Default::default()
+            }),
+        );
+        let env = generate_mcpg_docker_env(&fm);
+        assert!(
+            env.contains("-e AZURE_DEVOPS_EXT_PAT=\"$(SC_READ_TOKEN)\""),
+            "Should auto-map ADO token when permissions.read is set and MCP requests it"
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_docker_env_permissions_read_no_mcp_request() {
+        let mut fm = minimal_front_matter();
+        fm.permissions = Some(crate::compile::types::PermissionsConfig {
+            read: Some("my-read-sc".to_string()),
+            write: None,
+        });
+        // No MCP requests AZURE_DEVOPS_EXT_PAT — auto-map should NOT trigger
+        fm.mcp_servers.insert(
+            "unrelated-tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                ..Default::default()
+            }),
+        );
+        let env = generate_mcpg_docker_env(&fm);
+        assert!(
+            !env.contains("AZURE_DEVOPS_EXT_PAT"),
+            "Should NOT auto-map ADO token when no MCP requests it"
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_docker_env_dedup_auto_map_and_passthrough() {
+        // When permissions.read is set AND MCP has AZURE_DEVOPS_EXT_PAT: "",
+        // the auto-mapped form (with SC_READ_TOKEN) should win — no duplicate
+        let mut fm = minimal_front_matter();
+        fm.permissions = Some(crate::compile::types::PermissionsConfig {
+            read: Some("my-read-sc".to_string()),
+            write: None,
+        });
+        fm.mcp_servers.insert(
+            "ado-tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("node:20-slim".to_string()),
+                env: {
+                    let mut e = HashMap::new();
+                    e.insert("AZURE_DEVOPS_EXT_PAT".to_string(), "".to_string());
+                    e
+                },
+                ..Default::default()
+            }),
+        );
+        let env = generate_mcpg_docker_env(&fm);
+        // Should have the SC_READ_TOKEN form (auto-mapped), not bare passthrough
+        assert!(
+            env.contains("-e AZURE_DEVOPS_EXT_PAT=\"$(SC_READ_TOKEN)\""),
+            "Auto-mapped form should be present"
+        );
+        // Should appear exactly once
+        let count = env.matches("AZURE_DEVOPS_EXT_PAT").count();
+        assert_eq!(count, 1, "AZURE_DEVOPS_EXT_PAT should appear exactly once, got {}", count);
+    }
+
+    #[test]
+    fn test_generate_mcpg_docker_env_without_permissions() {
+        let fm = minimal_front_matter();
+        let env = generate_mcpg_docker_env(&fm);
+        assert!(
+            !env.contains("AZURE_DEVOPS_EXT_PAT"),
+            "Should not map ADO token when permissions.read is not set"
+        );
+    }
+
+    #[test]
+    fn test_generate_mcpg_docker_env_passthrough_vars() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "tool".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("img:latest".to_string()),
+                env: {
+                    let mut e = HashMap::new();
+                    e.insert("PASS_THROUGH".to_string(), "".to_string());
+                    e.insert("STATIC".to_string(), "value".to_string());
+                    e
+                },
+                ..Default::default()
+            }),
+        );
+        let env = generate_mcpg_docker_env(&fm);
+        assert!(env.contains("-e PASS_THROUGH"), "Should include passthrough var");
+        assert!(!env.contains("-e STATIC"), "Should NOT include static var");
+    }
+
+    #[test]
+    fn test_generate_mcpg_docker_env_rejects_invalid_names() {
+        let mut fm = minimal_front_matter();
+        fm.mcp_servers.insert(
+            "evil".to_string(),
+            McpConfig::WithOptions(McpOptions {
+                container: Some("img:latest".to_string()),
+                env: {
+                    let mut e = HashMap::new();
+                    // Injection attempt: env var name with Docker flag
+                    e.insert("MY_VAR --privileged".to_string(), "".to_string());
+                    // Valid env var for comparison
+                    e.insert("GOOD_VAR".to_string(), "".to_string());
+                    e
+                },
+                ..Default::default()
+            }),
+        );
+        let env = generate_mcpg_docker_env(&fm);
+        assert!(
+            !env.contains("--privileged"),
+            "Should reject invalid env var name with Docker flag injection"
+        );
+        assert!(
+            env.contains("-e GOOD_VAR"),
+            "Should include valid env var"
+        );
+    }
+
+    #[test]
+    fn test_is_valid_env_var_name() {
+        assert!(is_valid_env_var_name("MY_VAR"));
+        assert!(is_valid_env_var_name("_PRIVATE"));
+        assert!(is_valid_env_var_name("A"));
+        assert!(is_valid_env_var_name("VAR123"));
+        assert!(!is_valid_env_var_name(""));
+        assert!(!is_valid_env_var_name("123ABC"));
+        assert!(!is_valid_env_var_name("MY-VAR"));
+        assert!(!is_valid_env_var_name("MY VAR"));
+        assert!(!is_valid_env_var_name("X --privileged"));
+        assert!(!is_valid_env_var_name("X -v /etc:/etc:rw"));
+    }
+
+    // ─── tools.azure-devops MCPG integration ────────────────────────────────
+
+    #[test]
+    fn test_ado_tool_generates_mcpg_entry() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops: true\n---\n",
+        )
+        .unwrap();
+        // Pass inferred org since no explicit org is set
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test_with_org(&fm, "inferred-org"), &collect_extensions(&fm)).unwrap();
+        let ado = config.mcp_servers.get("azure-devops").unwrap();
+        assert_eq!(ado.server_type, "stdio");
+        assert_eq!(ado.container.as_deref(), Some(ADO_MCP_IMAGE));
+        assert_eq!(ado.entrypoint.as_deref(), Some(ADO_MCP_ENTRYPOINT));
+        let args = ado.entrypoint_args.as_ref().unwrap();
+        assert!(args.contains(&"-y".to_string()));
+        assert!(args.contains(&ADO_MCP_PACKAGE.to_string()));
+        assert!(args.contains(&"inferred-org".to_string()));
+        // Should have AZURE_DEVOPS_EXT_PAT in env
+        let env = ado.env.as_ref().unwrap();
+        assert!(env.contains_key("AZURE_DEVOPS_EXT_PAT"));
+    }
+
+    #[test]
+    fn test_ado_tool_with_toolsets() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    toolsets: [repos, wit, core]\n---\n",
+        )
+        .unwrap();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test_with_org(&fm, "myorg"), &collect_extensions(&fm)).unwrap();
+        let ado = config.mcp_servers.get("azure-devops").unwrap();
+        let args = ado.entrypoint_args.as_ref().unwrap();
+        assert!(args.contains(&"-d".to_string()));
+        assert!(args.contains(&"repos".to_string()));
+        assert!(args.contains(&"wit".to_string()));
+        assert!(args.contains(&"core".to_string()));
+    }
+
+    #[test]
+    fn test_ado_tool_with_org_override() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    org: myorg\n---\n",
+        )
+        .unwrap();
+        // Explicit org should be used even when inferred_org is None
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let ado = config.mcp_servers.get("azure-devops").unwrap();
+        let args = ado.entrypoint_args.as_ref().unwrap();
+        assert!(args.contains(&"myorg".to_string()));
+    }
+
+    #[test]
+    fn test_ado_tool_explicit_org_overrides_inferred() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    org: explicit-org\n---\n",
+        )
+        .unwrap();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test_with_org(&fm, "inferred-org"), &collect_extensions(&fm)).unwrap();
+        let ado = config.mcp_servers.get("azure-devops").unwrap();
+        let args = ado.entrypoint_args.as_ref().unwrap();
+        assert!(args.contains(&"explicit-org".to_string()));
+        assert!(!args.contains(&"inferred-org".to_string()));
+    }
+
+    #[test]
+    fn test_ado_tool_no_org_fails() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops: true\n---\n",
+        )
+        .unwrap();
+        // No explicit org and no inferred org — should fail
+        let result = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm));
+        assert!(result.is_err());
+        assert!(
+            result.unwrap_err().to_string().contains("no ADO organization"),
+            "Error should mention missing org"
+        );
+    }
+
+    #[test]
+    fn test_ado_tool_invalid_org_fails() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    org: \"my org/bad\"\n---\n",
+        )
+        .unwrap();
+        let result = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm));
+        assert!(result.is_err());
+        assert!(
+            result.unwrap_err().to_string().contains("Invalid ADO org name"),
+            "Error should mention invalid org"
+        );
+    }
+
+    #[test]
+    fn test_ado_tool_invalid_toolset_fails() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    org: myorg\n    toolsets: [\"repos\", \"bad toolset\"]\n---\n",
+        )
+        .unwrap();
+        let result = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm));
+        assert!(result.is_err());
+        assert!(
+            result.unwrap_err().to_string().contains("Invalid ADO toolset name"),
+            "Error should mention invalid toolset"
+        );
+    }
+
+    #[test]
+    fn test_ado_tool_with_allowed_tools() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    org: myorg\n    allowed:\n      - wit_get_work_item\n      - core_list_projects\n---\n",
+        )
+        .unwrap();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let ado = config.mcp_servers.get("azure-devops").unwrap();
+        let tools = ado.tools.as_ref().unwrap();
+        assert_eq!(tools, &["wit_get_work_item", "core_list_projects"]);
+    }
+
+    #[test]
+    fn test_ado_tool_disabled_not_generated() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops: false\n---\n",
+        )
+        .unwrap();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        assert!(!config.mcp_servers.contains_key("azure-devops"));
+    }
+
+    #[test]
+    fn test_ado_tool_not_set_not_generated() {
+        let fm = minimal_front_matter();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        assert!(!config.mcp_servers.contains_key("azure-devops"));
+    }
+
+    #[test]
+    fn test_ado_tool_skips_manual_mcp_entry() {
+        // When tools.azure-devops is enabled AND mcp-servers also has azure-devops,
+        // the tools config takes precedence and the manual entry is skipped.
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops:\n    org: auto-org\nmcp-servers:\n  azure-devops:\n    container: \"node:20-slim\"\n    entrypoint: \"npx\"\n    entrypoint-args: [\"-y\", \"@azure-devops/mcp\", \"manual-org\"]\n---\n",
+        )
+        .unwrap();
+        let config = generate_mcpg_config(&fm, &CompileContext::for_test(&fm), &collect_extensions(&fm)).unwrap();
+        let ado = config.mcp_servers.get("azure-devops").unwrap();
+        // Should use the auto-configured org, not the manual one
+        let args = ado.entrypoint_args.as_ref().unwrap();
+        assert!(args.contains(&"auto-org".to_string()));
+        assert!(!args.contains(&"manual-org".to_string()));
+    }
+
+    #[test]
+    fn test_ado_tool_docker_env_passthrough() {
+        let (fm, _) = parse_markdown(
+            "---\nname: test\ndescription: test\ntools:\n  azure-devops: true\npermissions:\n  read: my-read-sc\n---\n",
+        )
+        .unwrap();
+        let env = generate_mcpg_docker_env(&fm);
+        assert!(
+            env.contains("AZURE_DEVOPS_EXT_PAT"),
+            "Should include ADO token passthrough when permissions.read is set"
+        );
+    }
+
+    // ─── validate_docker_args ────────────────────────────────────────────────
+
+    #[test]
+    fn test_validate_docker_args_privileged_flag() {
+        let warnings = validate_docker_args(&["--privileged".to_string()], "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("--privileged"), "should warn about --privileged");
+    }
+
+    #[test]
+    fn test_validate_docker_args_entrypoint_in_args_warns() {
+        let warnings = validate_docker_args(
+            &[
+                "--entrypoint".to_string(),
+                "/bin/sh".to_string(),
+            ],
+            "my-mcp",
+        );
+        assert!(warnings.iter().any(|w| w.contains("--entrypoint") && w.contains("entrypoint:")),
+            "should warn about --entrypoint with hint to use entrypoint: field");
+    }
+
+    #[test]
+    fn test_validate_docker_args_volume_flag_calls_mount_validation() {
+        // -v docker.sock in args bypasses `mounts:` validation; should produce warnings
+        let warnings = validate_docker_args(
+            &[
+                "-v".to_string(),
+                "/var/run/docker.sock:/var/run/docker.sock".to_string(),
+            ],
+            "my-mcp",
+        );
+        assert!(warnings.iter().any(|w| w.contains("bypasses mounts validation")),
+            "should warn about volume mount in args");
+        assert!(warnings.iter().any(|w| w.contains("Docker socket")),
+            "should propagate mount source warning for docker.sock");
+    }
+
+    #[test]
+    fn test_validate_docker_args_volume_equals_form() {
+        // --volume=source:dest form should also be detected
+        let warnings = validate_docker_args(
+            &["--volume=/var/run/docker.sock:/var/run/docker.sock".to_string()],
+            "my-mcp",
+        );
+        assert!(warnings.iter().any(|w| w.contains("bypasses mounts validation")),
+            "should warn about --volume= form");
+    }
+
+    #[test]
+    fn test_validate_docker_args_safe_args_no_warnings() {
+        // A legitimate arg like --read-only should produce no warnings
+        let warnings = validate_docker_args(&["--read-only".to_string()], "my-mcp");
+        assert!(warnings.is_empty(), "safe args should not produce warnings");
+    }
+
+    #[test]
+    fn test_validate_docker_args_empty_no_warnings() {
+        let warnings = validate_docker_args(&[], "my-mcp");
+        assert!(warnings.is_empty(), "empty args should not produce warnings");
+    }
+
+    #[test]
+    fn test_validate_docker_args_volume_flag_trailing_warns() {
+        // -v as the last arg with no mount spec is malformed
+        let warnings = validate_docker_args(&["-v".to_string()], "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("malformed"), "trailing -v with no mount spec should warn");
+    }
+
+    #[test]
+    fn test_validate_docker_args_long_volume_flag_trailing_warns() {
+        // --volume as the last arg with no mount spec is malformed
+        let warnings = validate_docker_args(&["--volume".to_string()], "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("malformed"), "trailing --volume with no mount spec should warn");
+    }
+
+    // ─── validate_mcp_url ────────────────────────────────────────────────────
+
+    #[test]
+    fn test_validate_mcp_url_https_no_warnings() {
+        let warnings = validate_mcp_url("https://mcp.dev.azure.com/myorg", "my-mcp");
+        assert!(warnings.is_empty(), "https URL should not produce warnings");
+    }
+
+    #[test]
+    fn test_validate_mcp_url_http_no_warnings() {
+        let warnings = validate_mcp_url("http://localhost:8100/mcp", "my-mcp");
+        assert!(warnings.is_empty(), "http URL should not produce warnings");
+    }
+
+    #[test]
+    fn test_validate_mcp_url_bad_scheme_warns() {
+        let warnings = validate_mcp_url("ftp://files.example.com", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("does not use http://"), "non-HTTP scheme should warn");
+    }
+
+    #[test]
+    fn test_validate_mcp_url_no_scheme_warns() {
+        let warnings = validate_mcp_url("mcp.dev.azure.com/myorg", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("does not use http://"), "URL without scheme should warn");
+    }
+
+    // ─── validate_mount_source ───────────────────────────────────────────────
+
+    #[test]
+    fn test_validate_mount_source_docker_sock() {
+        let warnings = validate_mount_source("/var/run/docker.sock:/var/run/docker.sock:rw", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("Docker socket"), "should warn about Docker socket exposure");
+    }
+
+    #[test]
+    fn test_validate_mount_source_sensitive_path_etc() {
+        let warnings = validate_mount_source("/etc/passwd:/data/passwd:ro", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("sensitive host path"), "should warn about /etc mount");
+    }
+
+    #[test]
+    fn test_validate_mount_source_sensitive_path_proc() {
+        let warnings = validate_mount_source("/proc:/host/proc:ro", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("sensitive host path"), "should warn about /proc mount");
+    }
+
+    #[test]
+    fn test_validate_mount_source_case_insensitive() {
+        // /ETC/shadow should match sensitive /etc prefix (lowercased comparison)
+        let warnings = validate_mount_source("/ETC/shadow:/data/shadow:ro", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("sensitive host path"), "case-insensitive match should trigger warning");
+    }
+
+    #[test]
+    fn test_validate_mount_source_no_false_positive_on_etc_configs() {
+        // /etc-configs should NOT match the /etc prefix (path boundary check requires trailing /)
+        let warnings = validate_mount_source("/etc-configs:/app/config:ro", "my-mcp");
+        assert!(warnings.is_empty(), "/etc-configs must not match /etc prefix due to path boundary check");
+    }
+
+    #[test]
+    fn test_validate_mount_source_safe_path_no_warnings() {
+        // /app/data is not a sensitive path; should produce no warnings
+        let warnings = validate_mount_source("/app/data:/app/data:ro", "my-mcp");
+        assert!(warnings.is_empty(), "safe path should not produce warnings");
+    }
+
+    // ─── validate_container_image ────────────────────────────────────────────
+
+    #[test]
+    fn test_validate_container_image_empty_string() {
+        let warnings = validate_container_image("", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("empty"), "should warn about empty image name");
+    }
+
+    #[test]
+    fn test_validate_container_image_shell_metacharacters() {
+        let warnings = validate_container_image("node:20-slim; rm -rf /", "my-mcp");
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("unexpected characters"), "should warn about shell metacharacters");
+    }
+
+    #[test]
+    fn test_validate_container_image_valid_name_no_warnings() {
+        // Standard image references should produce no warnings
+        assert!(validate_container_image("node:20-slim", "my-mcp").is_empty());
+        assert!(validate_container_image("ghcr.io/org/image:latest", "my-mcp").is_empty());
+        assert!(validate_container_image("python:3.12-slim", "my-mcp").is_empty());
+    }
+
+    // ─── warn_potential_secrets ──────────────────────────────────────────────
+
+    #[test]
+    fn test_warn_potential_secrets_token_env_var_triggers() {
+        let env = HashMap::from([("API_TOKEN".to_string(), "secret123".to_string())]);
+        let headers = HashMap::new();
+        let warnings = warn_potential_secrets("my-mcp", &env, &headers);
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("API_TOKEN"), "should warn about secret-looking env var");
+    }
+
+    #[test]
+    fn test_warn_potential_secrets_empty_passthrough_no_warnings() {
+        // Empty string = passthrough; should NOT trigger a warning
+        let env = HashMap::from([("API_TOKEN".to_string(), "".to_string())]);
+        let headers = HashMap::new();
+        let warnings = warn_potential_secrets("my-mcp", &env, &headers);
+        assert!(warnings.is_empty(), "empty passthrough value must not trigger a warning");
+    }
+
+    #[test]
+    fn test_warn_potential_secrets_authorization_header_triggers() {
+        let env = HashMap::new();
+        let headers =
+            HashMap::from([("Authorization".to_string(), "Bearer abc".to_string())]);
+        let warnings = warn_potential_secrets("my-mcp", &env, &headers);
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("Authorization"), "should warn about Authorization header");
+    }
+
+    #[test]
+    fn test_warn_potential_secrets_bearer_value_triggers() {
+        // A header whose value starts with "Bearer " should also warn
+        let env = HashMap::new();
+        let headers =
+            HashMap::from([("X-Custom-Auth".to_string(), "Bearer token123".to_string())]);
+        let warnings = warn_potential_secrets("my-mcp", &env, &headers);
+        assert_eq!(warnings.len(), 1);
+        assert!(warnings[0].contains("X-Custom-Auth"), "should warn about header with Bearer value");
+    }
+
+    #[test]
+    fn test_warn_potential_secrets_safe_env_no_warnings() {
+        // Env keys with non-secret names and non-empty values should produce no warnings
+        let env = HashMap::from([("MY_CONFIG".to_string(), "value".to_string())]);
+        let headers = HashMap::new();
+        let warnings = warn_potential_secrets("my-mcp", &env, &headers);
+        assert!(warnings.is_empty(), "non-secret env var should not produce warnings");
     }
 }
